Massive update (bot api 6.3)

+++ src/tourmaline/models/audio.cr
@@ -17,7 +17,7 @@ module Tourmaline
 
     getter mime_type : String?
 
-    getter file_size : Int32?
+    getter file_size : Int64?
 
     getter thumb : PhotoSize?
   end
